clean code and DataGrid

+++ react-frontend/src/components/admin/DeleteCompany.tsx
@@ -5,47 +5,53 @@ import Card from '../UI/Card';
 import MessageModal from '../UI/MessageModal';
 import './Admin.css'
 
-type Props = {
 
-}
+const DeleteCompany = () => {
 
 
-const DeleteCompany: React.FC<Props> = (props: Props) => {
+    //company ID to update
+    const companyIdRef = useRef<HTMLInputElement>();
 
-    const companyIdRef = useRef<HTMLInputElement>();//company ID to update
-
+    //popup message
     const [messageState, setMessageState] = useState<{ title: string, message: string }>(null);
 
 
     const updateCompanyHandler = async (event) => {
+        event.preventDefault()
+
+        //check input
+        if (companyIdRef.current.value.trim().length === 0) {
+            setMessageState({ title: "Invalid input", message: "Please enter a valid inputs value! (No empty values)" })
+            return;
+        }
+
+
         try {
-
-            event.preventDefault()
-
-            console.log("Company ID to delete: " + companyIdRef.current.value)
-
-
-
-
             const requestOptions =
             {
                 method: 'DELETE',
                 headers: { "Content-Type": "application/json" },
             }
 
+            //java server side
             const response = await fetch("http://localhost:8080/CouponApp/deleteCompanyById/" + companyIdRef.current.value, requestOptions);
 
+            //using the first response from the server - to text
             const responseFromUpdateCompany = await response.text();
 
             if (response.status === 400) { //BAD REQUEST
-                console.log("Company by ID given to delete: " + companyIdRef.current.value)
                 setMessageState({ title: "Oops! something went wrong!", message: responseFromUpdateCompany })
+
+                // restart values
+                companyIdRef.current.value = ''
+
                 //response from controller
                 throw new Error(responseFromUpdateCompany)
             }
 
+            //if success (response status !== 400)
             setMessageState({ title: "Company has been deleted!", message: "Company has been deleted from the system successfully!" })
-            console.log(responseFromUpdateCompany);
+
 
             // restart values
             companyIdRef.current.value = ''
@@ -73,7 +79,7 @@ const DeleteCompany: React.FC<Props> = (props: Props) => {
                 <form onSubmit={updateCompanyHandler}>
 
                     <label className='label'>Company ID</label>
-                    <input className='input' type="text" ref={companyIdRef} />
+                    <input className='input' type="number" ref={companyIdRef} />
                     <br />
                     <div>
                         <Button type="submit">Delete company</Button>
